Ch. 17 §02: correct numbering and content for Listings 17-08 and 17-09
