+++ listings/ch17-async-await/listing-17-09/src/main.rs
@@ -1,46 +1,13 @@
-use std::time::Duration;
-
 fn main() {
     trpl::block_on(async {
-        // ANCHOR: here
         let (tx, mut rx) = trpl::channel();
 
-        let tx1 = tx.clone();
-        let tx1_fut = async move {
-            let vals = vec![
-                String::from("hi"),
-                String::from("from"),
-                String::from("the"),
-                String::from("future"),
-            ];
+        // ANCHOR: send-and-receive
+        let val = String::from("hi");
+        tx.send(val).unwrap();
 
-            for val in vals {
-                tx1.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        };
-
-        let rx_fut = async {
-            while let Some(value) = rx.recv().await {
-                println!("received '{value}'");
-            }
-        };
-
-        let tx_fut = async move {
-            let vals = vec![
-                String::from("more"),
-                String::from("messages"),
-                String::from("for"),
-                String::from("you"),
-            ];
-
-            for val in vals {
-                tx.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        };
-
-        trpl::join3(tx1_fut, tx_fut, rx_fut).await;
-        // ANCHOR_END: here
+        let received = rx.recv().await.unwrap();
+        println!("Got: {received}");
+        // ANCHOR_END: send-and-receive
     });
 }
